+++ mat_dp_pipeline/data_sources/__init__.py
@@ -1,9 +1,16 @@
-from mat_dp_pipeline.data_sources.generic import (
+from mat_dp_pipeline.data_sources.definitions import (
     BaseSource,
     IndicatorsSource,
     IntensitiesSource,
+    TargetsSource,
+)
+from mat_dp_pipeline.data_sources.mat_dp_db import (
+    MatDPDBIndicatorsSource,
+    MatDPDBIntensitiesSource,
+)
+from mat_dp_pipeline.data_sources.stored import (
     StoredIndicators,
     StoredIntensities,
     StoredTargets,
-    TargetsSource,
 )
+from mat_dp_pipeline.data_sources.tmba import TMBATargetsSource
